// apps/media-live-service/src/modules/live/enums/live-status.enum.ts
export enum LiveStatus {
  WAITING = 'WAITING',
  LIVE = 'LIVE',
  ENDED = 'ENDED',
  LOCKED = 'LOCKED',
}

export enum LiveCategory {
  EXCLUSIVE = "EXCLUSIVE",
  CERTIFIED = "CERTIFIED",
  WOMAN = "WOMAN",
  HOT_SEDUCTION = "HOT_SEDUCTION",
  SOULMATE = "SOULMATE",
  MATURE = "MATURE",
  NEW_MODELS = "NEW_MODELS",
  FETISH = "FETISH",
  TRANS = "TRANS",
  LESBIANS = "LESBIANS",
  COUPLES = "COUPLES",
  SPREAD_PUSSY = "SPREAD_PUSSY",
  TOPLESS = "TOPLESS",
  SEX_TOY = "SEX_TOY",
  FINGERING = "FINGERING",
  ANAL = "ANAL",
  MASTURBATION = "MASTURBATION",
  ASS_PLAY = "ASS_PLAY",
  CLOSE_UP = "CLOSE_UP",
  BAREFOOT = "BAREFOOT",
  HEELS_ON = "HEELS_ON",
  SPREADING_LEGS = "SPREADING_LEGS",
  LYING = "LYING",
  IN_STOCKINGS = "IN_STOCKINGS",
  TATTOO = "TATTOO",
  LINGERIE = "LINGERIE",
  GLASSES_ON = "GLASSES_ON",
  STANDING = "STANDING",
  CROSSED_LEGS = "CROSSED_LEGS",
  BIRTHMARK = "BIRTHMARK"
}